+++ components/src/Button/IconicButton.js
@@ -2,10 +2,24 @@ import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 import { space } from "styled-system";
+import { Manager, Reference, Popper } from "react-popper";
+import { transparentize } from "polished";
 import { Icon } from "../Icon";
 import { Text } from "../Type";
 import theme from "../theme";
 import icons from "../../icons/icons.json";
+
+const HoverText = styled.div({
+  whiteSpace: "nowrap",
+  fontSize: theme.fontSizes.small,
+  lineHeight: theme.lineHeights.smallTextCompressed,
+  color: theme.colors.whiteGrey,
+  backgroundColor: transparentize(0.15, theme.colors.blackBlue),
+  borderRadius: theme.radii.medium,
+  marginTop: theme.space.half,
+  padding: `${theme.space.half} ${theme.space.x1}`,
+  pointerEvents: "none"
+});
 
 const WrapperButton = styled.button(space, ({ disabled }) => ({
   background: "transparent",
@@ -26,9 +40,15 @@ const WrapperButton = styled.button(space, ({ disabled }) => ({
     fontWeight: theme.fontWeights.medium,
     textAlign: "left"
   },
+  [`${HoverText}`]: {
+    opacity: "0"
+  },
   "&:hover": {
     [`${Icon}`]: {
       backgroundColor: theme.colors.lightBlue
+    },
+    [`${HoverText}`]: {
+      opacity: "1"
     }
   },
   "&:active": {
@@ -50,38 +70,50 @@ const WrapperButton = styled.button(space, ({ disabled }) => ({
     outline: "none",
     [`${Icon}`]: {
       boxShadow: theme.shadows.focus
+    },
+    [`${HoverText}`]: {
+      opacity: "1"
     }
   }
 }));
 
-const BaseIconicButton = React.forwardRef((props, ref) => {
-  const { children, icon } = props;
-
-  return (
-    <WrapperButton ref={ref} label={children} {...props}>
-      <Icon size={theme.space.x4} icon={icon} p="half" />
-      {children && (
-        <Text mr="half" ml="half">
-          {children}
-        </Text>
-      )}
-    </WrapperButton>
-  );
-});
+const BaseIconicButton = React.forwardRef(({ children, icon, labelHidden, ...props }, forwardedRef) => (
+  <WrapperButton ref={forwardedRef} label={children} {...props}>
+    <Manager>
+      <Reference>{({ ref }) => <Icon ref={ref} size={theme.space.x4} icon={icon} p="half" />}</Reference>
+      <Popper placement="bottom">
+        {({ ref, style, placement }) =>
+          labelHidden ? (
+            <HoverText ref={ref} style={style} placement={placement}>
+              {children}
+            </HoverText>
+          ) : null
+        }
+      </Popper>
+    </Manager>
+    {children && !labelHidden && (
+      <Text mr="half" ml="half">
+        {children}
+      </Text>
+    )}
+  </WrapperButton>
+));
 
 const IconicButton = styled(BaseIconicButton)({});
 
 export const iconNames = Object.keys(icons);
 
 BaseIconicButton.propTypes = {
-  children: PropTypes.node,
+  children: PropTypes.string,
+  labelHidden: PropTypes.bool,
   disabled: PropTypes.bool,
   icon: PropTypes.oneOf(iconNames).isRequired
 };
 
 BaseIconicButton.defaultProps = {
-  disabled: false,
-  children: null
+  children: null,
+  labelHidden: false,
+  disabled: false
 };
 
 export default IconicButton;