+++ components/src/Button/IconicButton.story.js
@@ -3,13 +3,11 @@ import { storiesOf } from "@storybook/react";
 import { IconicButton } from "../index";
 
 storiesOf("IconicButton", module)
-  .add("With label", () => <IconicButton icon="delete">Delete</IconicButton>)
-  .add("With a long label", () => (
-    <React.Fragment>
-      <IconicButton icon="user">I am an Iconic Button with a really really really long label</IconicButton>
-    </React.Fragment>
+  .add("with label", () => <IconicButton icon="delete">Delete</IconicButton>)
+  .add("with a long label", () => (
+    <IconicButton icon="user">I am an Iconic Button with a really really really long label</IconicButton>
   ))
-  .add("Disabled", () => (
+  .add("set to disabled", () => (
     <React.Fragment>
       <IconicButton icon="cancel" disabled>
         Cancel
@@ -18,4 +16,9 @@ storiesOf("IconicButton", module)
         Lock
       </IconicButton>
     </React.Fragment>
+  ))
+  .add("with a hidden label", () => (
+    <IconicButton ml="x6" labelHidden icon="user">
+      Hidden Label
+    </IconicButton>
   ));